terraform-provider-glesys
Provide functions to create some of the basic functions provided by the
GleSYS api.

Resources:
Server - create/read/update/delete a vps instance.
Network - create networks in the datacenters
Networkadapter - interfaces to be used with server and network

TODO:
Tests - for each of the resources
Minor bugs when deleting networks, fix the lifecycle so that the network
isn't deleted before the attached network interface.